chore(signature-model): remove pydantic and attrs signature models
