+++ .pre-commit-config.yaml
@@ -76,7 +76,6 @@ repos:
             beanie,
             beautifulsoup4,
             brotli,
-            cattrs,
             click,
             fakeredis>=2.10.2,
             fast-query-parsers,
@@ -136,7 +135,6 @@ repos:
             beanie,
             beautifulsoup4,
             brotli,
-            cattrs,
             click,
             fakeredis>=2.10.2,
             fast-query-parsers,
