+++ litestar/_signature/__init__.py
@@ -1,4 +1,4 @@
-from .models.base import SignatureModel
-from .utils import create_signature_model, get_signature_model
+from .model import SignatureModel
+from .utils import get_signature_model
 
-__all__ = ("create_signature_model", "SignatureModel", "get_signature_model")
+__all__ = ("SignatureModel", "get_signature_model")
