--- litestar/_signature/models/__init__.py
@@ -1,3 +1,0 @@
-from .base import SignatureModel
-
-__all__ = ("SignatureModel",)
