+++ tests/unit/test_dto/test_integration.py
@@ -15,7 +15,7 @@ def test_dto_defined_on_handler() -> None:
         assert data == Model(a=1, b="2")
         return data
 
-    with create_test_client(route_handlers=handler, _preferred_validation_backend="pydantic") as client:
+    with create_test_client(route_handlers=handler) as client:
         response = client.post("/", json={"what": "ever"})
         assert response.status_code == 201
         assert response.json() == {"a": 1, "b": "2"}
