+++ tests/unit/test_openapi/test_parameters.py
@@ -7,7 +7,7 @@ from litestar import Controller, Litestar, Router, get
 from litestar._openapi.parameters import create_parameter_for_handler
 from litestar._openapi.schema_generation import SchemaCreator
 from litestar._openapi.typescript_converter.schema_parsing import is_schema_value
-from litestar._signature import create_signature_model
+from litestar._signature import SignatureModel
 from litestar.di import Provide
 from litestar.enums import ParamType
 from litestar.exceptions import ImproperlyConfiguredException
@@ -28,10 +28,9 @@ def _create_parameters(app: Litestar, path: str) -> List["OpenAPIParameter"]:
     handler = route_handler.fn.value
     assert callable(handler)
 
-    handler_fields = create_signature_model(
+    handler_fields = SignatureModel.create(
         fn=handler,
         dependency_name_set=set(),
-        preferred_validation_backend=app._preferred_validation_backend,
         parsed_signature=route_handler.parsed_fn_signature,
     ).fields
