+++ tests/unit/test_signature/test_validation.py
@@ -1,0 +1,319 @@
+from dataclasses import dataclass
+from typing import List, Optional
+
+import pytest
+from attr import define
+from pydantic import BaseModel
+from typing_extensions import TypedDict
+
+from litestar import get, post
+from litestar._signature import SignatureModel
+from litestar.di import Provide
+from litestar.exceptions import ImproperlyConfiguredException, ValidationException
+from litestar.params import Dependency, Parameter
+from litestar.status_codes import HTTP_400_BAD_REQUEST, HTTP_500_INTERNAL_SERVER_ERROR
+from litestar.testing import RequestFactory, create_test_client
+from litestar.utils.signature import ParsedSignature
+
+
+def test_parses_values_from_connection_kwargs_raises() -> None:
+    def fn(a: int) -> None:
+        pass
+
+    model = SignatureModel.create(
+        fn=fn,
+        dependency_name_set=set(),
+        parsed_signature=ParsedSignature.from_fn(fn, {}),
+    )
+    with pytest.raises(ValidationException):
+        model.parse_values_from_connection_kwargs(connection=RequestFactory().get(), a="not an int")
+
+
+def test_create_signature_validation() -> None:
+    @get()
+    def my_fn(typed: int, untyped) -> None:  # type: ignore
+        pass
+
+    with pytest.raises(ImproperlyConfiguredException):
+        SignatureModel.create(
+            fn=my_fn.fn.value,
+            dependency_name_set=set(),
+            parsed_signature=ParsedSignature.from_fn(my_fn.fn.value, {}),
+        )
+
+
+def test_dependency_validation_failure_raises_500() -> None:
+    dependencies = {"dep": Provide(lambda: "thirteen", sync_to_thread=False)}
+
+    @get("/")
+    def test(dep: int, param: int, optional_dep: Optional[int] = Dependency()) -> None:
+        ...
+
+    with create_test_client(
+        route_handlers=[test],
+        dependencies=dependencies,
+    ) as client:
+        response = client.get("/?param=13")
+
+    assert response.json() == {"detail": "Internal Server Error", "status_code": HTTP_500_INTERNAL_SERVER_ERROR}
+
+
+def test_validation_failure_raises_400() -> None:
+    dependencies = {"dep": Provide(lambda: 13, sync_to_thread=False)}
+
+    @get("/")
+    def test(dep: int, param: int, optional_dep: Optional[int] = Dependency()) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test], dependencies=dependencies) as client:
+        response = client.get("/?param=thirteen")
+
+    assert response.json() == {
+        "detail": "Validation failed for GET http://testserver.local/?param=thirteen",
+        "extra": [{"key": "param", "message": "Expected `int`, got `str`", "source": "query"}],
+        "status_code": 400,
+    }
+
+
+def test_client_backend_error_precedence_over_server_error() -> None:
+    dependencies = {
+        "dep": Provide(lambda: "thirteen", sync_to_thread=False),
+        "optional_dep": Provide(lambda: "thirty-one", sync_to_thread=False),
+    }
+
+    @get("/")
+    def test(dep: int, param: int, optional_dep: Optional[int] = Dependency()) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test], dependencies=dependencies) as client:
+        response = client.get("/?param=thirteen")
+
+    assert response.json() == {
+        "detail": "Validation failed for GET http://testserver.local/?param=thirteen",
+        "extra": [{"key": "param", "message": "Expected `int`, got `str`", "source": "query"}],
+        "status_code": 400,
+    }
+
+
+def test_validation_error_exception_key() -> None:
+    class OtherChild(BaseModel):
+        val: List[int]
+
+    class Child(BaseModel):
+        val: int
+        other_val: int
+
+    class Parent(BaseModel):
+        child: Child
+        other_child: OtherChild
+
+    def fn(data: Parent) -> None:
+        pass
+
+    model = SignatureModel.create(
+        fn=fn,
+        dependency_name_set=set(),
+        parsed_signature=ParsedSignature.from_fn(fn, {}),
+    )
+
+    with pytest.raises(ValidationException) as exc_info:
+        model.parse_values_from_connection_kwargs(
+            connection=RequestFactory().get(), data={"child": {}, "other_child": {}}
+        )
+
+    assert isinstance(exc_info.value.extra, list)
+    assert exc_info.value.extra[0]["key"] == "child.val"
+    assert exc_info.value.extra[1]["key"] == "child.other_val"
+    assert exc_info.value.extra[2]["key"] == "other_child.val"
+
+
+def test_invalid_input_pydantic() -> None:
+    class OtherChild(BaseModel):
+        val: List[int]
+
+    class Child(BaseModel):
+        val: int
+        other_val: int
+
+    class Parent(BaseModel):
+        child: Child
+        other_child: OtherChild
+
+    @post("/")
+    def test(
+        data: Parent,
+        int_param: int,
+        length_param: str = Parameter(min_length=2),
+        int_header: int = Parameter(header="X-SOME-INT"),
+        int_cookie: int = Parameter(cookie="int-cookie"),
+    ) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test]) as client:
+        response = client.post(
+            "/",
+            json={"child": {"val": "a", "other_val": "b"}, "other_child": {"val": [1, "c"]}},
+            params={"int_param": "param", "length_param": "d"},
+            headers={"X-SOME-INT": "header"},
+            cookies={"int-cookie": "cookie"},
+        )
+
+        assert response.status_code == HTTP_400_BAD_REQUEST
+
+        data = response.json()
+
+        assert data
+        assert data["extra"] == [
+            {"key": "child.val", "message": "value is not a valid integer", "source": "body"},
+            {"key": "child.other_val", "message": "value is not a valid integer", "source": "body"},
+            {"key": "other_child.val.1", "message": "value is not a valid integer", "source": "body"},
+            {"key": "int_param", "message": "value is not a valid integer", "source": "query"},
+            {"key": "length_param", "message": "ensure this value has at least 2 characters", "source": "query"},
+            {"key": "int_header", "message": "value is not a valid integer", "source": "header"},
+            {"key": "int_cookie", "message": "value is not a valid integer", "source": "cookie"},
+        ]
+
+
+def test_invalid_input_attrs() -> None:
+    @define
+    class OtherChild:
+        val: List[int]
+
+    @define
+    class Child:
+        val: int
+        other_val: int
+
+    @define
+    class Parent:
+        child: Child
+        other_child: OtherChild
+
+    @post("/")
+    def test(
+        data: Parent,
+        int_param: int,
+        int_header: int = Parameter(header="X-SOME-INT"),
+        int_cookie: int = Parameter(cookie="int-cookie"),
+    ) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test]) as client:
+        response = client.post(
+            "/",
+            json={"child": {"val": "a", "other_val": "b"}, "other_child": {"val": [1, "c"]}},
+            params={"int_param": "param"},
+            headers={"X-SOME-INT": "header"},
+            cookies={"int-cookie": "cookie"},
+        )
+
+        assert response.status_code == HTTP_400_BAD_REQUEST
+
+        data = response.json()
+
+        assert data
+        assert data["extra"] == [
+            {"key": "child.val", "message": "invalid literal for int() with base 10: 'a'", "source": "body"},
+            {"key": "child.other_val", "message": "invalid literal for int() with base 10: 'b'", "source": "body"},
+            {"key": "other_child.val.1", "message": "invalid literal for int() with base 10: 'c'", "source": "body"},
+            {"key": "int_param", "message": "invalid literal for int() with base 10: 'param'", "source": "query"},
+            {"key": "int_header", "message": "invalid literal for int() with base 10: 'header'", "source": "header"},
+            {"key": "int_cookie", "message": "invalid literal for int() with base 10: 'cookie'", "source": "cookie"},
+        ]
+
+
+def test_invalid_input_dataclass() -> None:
+    @dataclass
+    class OtherChild:
+        val: List[int]
+
+    @dataclass
+    class Child:
+        val: int
+        other_val: int
+
+    @dataclass
+    class Parent:
+        child: Child
+        other_child: OtherChild
+
+    @post("/")
+    def test(
+        data: Parent,
+        int_param: int,
+        length_param: str = Parameter(min_length=2),
+        int_header: int = Parameter(header="X-SOME-INT"),
+        int_cookie: int = Parameter(cookie="int-cookie"),
+    ) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test]) as client:
+        response = client.post(
+            "/",
+            json={"child": {"val": "a", "other_val": "b"}, "other_child": {"val": [1, "c"]}},
+            params={"int_param": "param", "length_param": "d"},
+            headers={"X-SOME-INT": "header"},
+            cookies={"int-cookie": "cookie"},
+        )
+
+        assert response.status_code == HTTP_400_BAD_REQUEST
+
+        data = response.json()
+
+        assert data
+        assert data["extra"] == [
+            {"key": "child.val", "message": "value is not a valid integer", "source": "body"},
+            {"key": "child.other_val", "message": "value is not a valid integer", "source": "body"},
+            {"key": "other_child.val.1", "message": "value is not a valid integer", "source": "body"},
+            {"key": "int_param", "message": "value is not a valid integer", "source": "query"},
+            {"key": "length_param", "message": "ensure this value has at least 2 characters", "source": "query"},
+            {"key": "int_header", "message": "value is not a valid integer", "source": "header"},
+            {"key": "int_cookie", "message": "value is not a valid integer", "source": "cookie"},
+        ]
+
+
+def test_invalid_input_typed_dict() -> None:
+    class OtherChild(TypedDict):
+        val: List[int]
+
+    class Child(TypedDict):
+        val: int
+        other_val: int
+
+    class Parent(TypedDict):
+        child: Child
+        other_child: OtherChild
+
+    @post("/")
+    def test(
+        data: Parent,
+        int_param: int,
+        length_param: str = Parameter(min_length=2),
+        int_header: int = Parameter(header="X-SOME-INT"),
+        int_cookie: int = Parameter(cookie="int-cookie"),
+    ) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test]) as client:
+        response = client.post(
+            "/",
+            json={"child": {"val": "a", "other_val": "b"}, "other_child": {"val": [1, "c"]}},
+            params={"int_param": "param", "length_param": "d"},
+            headers={"X-SOME-INT": "header"},
+            cookies={"int-cookie": "cookie"},
+        )
+
+        assert response.status_code == HTTP_400_BAD_REQUEST
+
+        data = response.json()
+
+        assert data
+        assert data["extra"] == [
+            {"key": "child.val", "message": "value is not a valid integer", "source": "body"},
+            {"key": "child.other_val", "message": "value is not a valid integer", "source": "body"},
+            {"key": "other_child.val.1", "message": "value is not a valid integer", "source": "body"},
+            {"key": "int_param", "message": "value is not a valid integer", "source": "query"},
+            {"key": "length_param", "message": "ensure this value has at least 2 characters", "source": "query"},
+            {"key": "int_header", "message": "value is not a valid integer", "source": "header"},
+            {"key": "int_cookie", "message": "value is not a valid integer", "source": "cookie"},
+        ]
